Overcoming issues with submodules and svn:externals, updated task and README to reflect
svn:externals cannot be setup for files that already exist - as the svn
bridge from github includes stub folders for submodules in lib/vendor,
we have to use the project lib/vendor folder to store the JS externals.
The task now checks to see if the project's lib/vendor folder contains
the files, or the plugin's.